March 14th end of the day

+++ ft_base_function.c
@@ -6,7 +6,7 @@
 /*   By: shabibol <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/10 20:11:37 by shabibol          #+#    #+#             */
-/*   Updated: 2022/03/14 18:15:03 by shabibol         ###   ########.fr       */
+/*   Updated: 2022/03/14 19:53:51 by shabibol         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft_printf.h"
@@ -48,7 +48,7 @@ int	ft_atoi(const char *str)
 	return (result * sign);
 }
 
-int	ft_get_digits(int n)
+int	ft_get_digits(long int n)
 {
 	int	size;
 
@@ -61,11 +61,11 @@ int	ft_get_digits(int n)
 	return (size);
 }
 
-char	*ft_itoa(int n)
+char	*ft_itoa(long int n)
 {
 	char		*result;
 	long int	num;
-	size_t		digit;
+	long int	digit;
 
 	num = n;
 	digit = ft_get_digits(n);
@@ -88,7 +88,7 @@ char	*ft_itoa(int n)
 	return (result);
 }
 
-char	*ft_itoa_sign(int n)
+char	*ft_itoa_sign(long int n)
 {
 	char		*result;
 	long int	num;
